14/2/25 video-22

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -39,26 +39,22 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-        try {
+        
 
             //policy
-            $this->isAble('store', Ticket::class);
+            if ($this->isAble('store', Ticket::class)) {
 
-            return new TicketResource(Ticket::create($request->mappedAttributes()));
-        } catch (AuthorizationException $ex) {
+                return new TicketResource(Ticket::create($request->mappedAttributes()));
+            }
+      
             return $this->error('You are not authorized to store the resource', 401);
-        }
+        
     }
 
 
 
 
-    /**
-     * Undocumented function
-     *
-     * @param [type] $ticket_id
-     * @return void
-     */
+  
     public function show($ticket_id)
     {
         try {
@@ -67,7 +63,7 @@ class TicketController extends ApiController
                 return new TicketResource($ticket->load('user'));
             }
             return new TicketResource($ticket);
-        } catch (AuthorizationException $exception) {
+        } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found', 404);
         }
     }
@@ -79,16 +75,14 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
 
-
-            $ticket->update($request->mappedAttributes());
-
-            return new TicketResource($ticket);
+                return new TicketResource($ticket);
+            }
+            return $this->error('You are not authorized to update the resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        }catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update the resource', 401);
         }
     }
 
@@ -108,11 +102,11 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             //policy
-            $this->isAble('replace', null);
+            if ($this->isAble('replace', null)) {
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
-
-            return new TicketResource($ticket);
+                return new TicketResource($ticket);
+            }
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
         }
@@ -132,11 +126,14 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
             //policy
-            $this->isAble('delete', null);
+            if ($this->isAble('delete', null)) {
 
-            $ticket->delete();
+                $ticket->delete();
 
-            return $this->ok('Ticket successfully deleted');
+                return $this->ok('Ticket successfully deleted');
+            }
+
+            return $this->error('You are not authorized to delete that user.', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket not found', 404);
         }
